Use MSRV for Clippy check
ccb8e23 bumped the Rust version used in the release process to follow
the stable channel, but left CI pinned to a specific nightly
release. This updates the Rust version, but rather than updating the
CI job to also follow the stable channel, this uses the MSRV specified
in the manifest. This will allow us to control when the Clippy lints
change rather than them changing whenever there is a new stable
release.