Use newer GitHub action for Clippy check
This might later prove to be a mistake, given the close relationship
this action has with its predecessor, but this at least supports
reporting job failure when Clippy fails. The token, with scoped write
permission, is needed to show the results (via annotations) when the
pull request is from a local branch. Without it, the results are
burried in the log output. Pull requests from a remote branch only
ever get read-only tokens, so the annotations are skipped by
action-rs/clippy-check and the results are instead highlighted in the
log output.