Update dependencies to address GHSA-g98v-hv3f-hcfr
atty appears to be abandoned and has been removed in recent versions
of clap and pretty_env_logger. In order to avoid bumping the MSRV,
clap is being held back to a version before 4.4, where its MSRV was
bumped to 1.70. A couple of transitive dependencies also needed to be
explicitly downgraded:

    cargo update -p anstyle@1.0.3 --precise 1.0.2
    cargo update -p clap_lex@0.5.1 --precise 0.5.0